Moved TextureAtlas to common directory, added conditionals based on renderer type so now the same atlas class is used in both renderers
