--- src/renderer/GPURenderer/Desktop/TextureAtlas/constants.js
@@ -1,1 +1,0 @@
-export const DATA_TEXTURE_SIZE = 256;
