--- src/renderer/GPURenderer/Desktop/shaders/constants.js
@@ -1,1 +1,0 @@
-export const SIZE_ATTENUATION_FACTOR = '600.0';
