--- src/renderer/GPURenderer/Mobile/TextureAtlas/constants.js
@@ -1,1 +1,0 @@
-export const DATA_TEXTURE_SIZE = 256;
