+++ src/renderer/GPURenderer/common/TextureAtlas/index.js
@@ -1,3 +1,8 @@
+import {
+  RENDERER_TYPE_GPU_DESKTOP,
+  RENDERER_TYPE_GPU_MOBILE,
+} from '../../../types';
+
 import { DATA_TEXTURE_SIZE } from './constants';
 import { __DEV__ } from '../../../../constants';
 import potpack from 'potpack';
@@ -8,13 +13,25 @@ import potpack from 'potpack';
  */
 export default class TextureAtlas {
   constructor(renderer, debug) {
-    const { three: THREE } = renderer;
+    const { three: THREE, type: rendererType } = renderer;
+    const data = new Float32Array(DATA_TEXTURE_SIZE * 4);
     const ctx = (this.ctx = document.createElement('canvas').getContext('2d'));
     const { canvas } = ctx;
 
-    this.indexData = new Float32Array(DATA_TEXTURE_SIZE * 4);
+    this.rendererType = rendererType;
+    this.indexData = data;
     this.canvas = canvas;
     this.entries = [];
+
+    if (rendererType === RENDERER_TYPE_GPU_DESKTOP) {
+      this.atlasIndex = new THREE.DataTexture(
+        data,
+        DATA_TEXTURE_SIZE,
+        1,
+        THREE.RGBAFormat,
+        THREE.FloatType
+      );
+    }
 
     canvas.width = canvas.height = DATA_TEXTURE_SIZE;
 
@@ -26,6 +43,11 @@ export default class TextureAtlas {
     this.atlasTexture.flipY = false;
 
     renderer.material.uniforms.uTexture.value = this.atlasTexture;
+
+    if (rendererType === RENDERER_TYPE_GPU_DESKTOP) {
+      renderer.material.uniforms.atlasIndex.value = this.atlasIndex;
+    }
+
     renderer.material.uniformsNeedUpdate = true;
   }
 
@@ -92,7 +114,15 @@ export default class TextureAtlas {
       return;
     }
 
-    const { entries, canvas, indexData, ctx, atlasTexture } = this;
+    const {
+      entries,
+      canvas,
+      indexData,
+      ctx,
+      atlasIndex,
+      atlasTexture,
+      rendererType,
+    } = this;
 
     for (let i = 0; i < entries.length; i++) {
       if (!entries[i].texture.image) {
@@ -123,12 +153,25 @@ export default class TextureAtlas {
     for (let i = 0, ii = 0; i < entries.length; i++, ii += 4) {
       const e = this.entries[i];
 
-      indexData[ii + 0] = e.x / (canvas.width + 1);
-      indexData[ii + 1] = e.y / (canvas.height + 1);
-      indexData[ii + 2] = (e.x + e.w) / (canvas.width + 1);
-      indexData[ii + 3] = (e.y + e.h) / (canvas.height + 1);
+      if (rendererType === RENDERER_TYPE_GPU_DESKTOP) {
+        indexData[ii + 0] = e.x / canvas.width;
+        indexData[ii + 1] = e.y / canvas.height;
+        indexData[ii + 2] = (e.x + e.w) / canvas.width;
+        indexData[ii + 3] = (e.y + e.h) / canvas.height;
+      }
+
+      if (rendererType === RENDERER_TYPE_GPU_MOBILE) {
+        indexData[ii + 0] = e.x / (canvas.width + 1);
+        indexData[ii + 1] = e.y / (canvas.height + 1);
+        indexData[ii + 2] = (e.x + e.w) / (canvas.width + 1);
+        indexData[ii + 3] = (e.y + e.h) / (canvas.height + 1);
+      }
 
       ctx.drawImage(e.texture.image, e.x, e.y, e.w, e.h);
+    }
+
+    if (rendererType === RENDERER_TYPE_GPU_DESKTOP) {
+      atlasIndex.needsUpdate = true;
     }
 
     atlasTexture.needsUpdate = true;
