+++ src/renderer/GPURenderer/common/index.js
@@ -1,2 +1,3 @@
 export { Target, UniqueList } from './stores';
 export { default as ParticleBuffer } from './ParticleBuffer';
+export { default as TextureAtlas } from './TextureAtlas';
