--- src/renderer/GPURenderer/shaders/index.js
@@ -1,2 +1,0 @@
-export { fragmentShader } from './fragmentShader';
-export { vertexShader } from './vertexShader';
