--- src/renderer/GPURenderer/shaders/utils.js
@@ -1,5 +1,0 @@
-export const withoutTextureAtlas = shader =>
-  shader
-    .split('\n')
-    .filter(e => !e.endsWith('//GPU'))
-    .join('\n');
